Combined `fetchUserData` and `updateUserData` URLs into a single `userData` URL
