+++ lib/nametag_db.ts
@@ -7,7 +7,7 @@ export async function fetchNametagFromDB(): Promise<NameTagContent | undefined> 
     if (session && session.user) {
         let nameTagToReturn = undefined;
 
-        await fetch("/api/auth/users/fetchUserData/nameTag?userEmail=" + session.user.email, { method: "GET" })
+        await fetch("/api/auth/users/userData/nameTag?userEmail=" + session.user.email, { method: "GET" })
         .then((res) => res.json())
         .then((resJson) => {
             if (resJson.success && resJson.nameTag) {
@@ -33,7 +33,7 @@ export async function updateNameTagInDB(newNameTag: NameTagContent) {
     const session = await getSession();
 
     if (session && session.user) {
-        await fetch("/api/auth/users/updateUserData/nameTag", { 
+        await fetch("/api/auth/users/userData/nameTag", { 
             method: "POST",
             body: JSON.stringify({
                 email: session.user.email,
